+++ src/app/admin/path/path-list.component.ts
