+++ src/app/shared/services/user/user-title-resolver.service.ts
@@ -3,7 +3,7 @@ import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router';
 
 import { map } from 'rxjs';
 
-import { UsersService } from '@services/user.service';
+import { UsersService } from '@services/user/user.service';
 
 export const userNameResolver: ResolveFn<string> = (route: ActivatedRouteSnapshot) => {
   const id = route.paramMap.get('id');
