+++ tsconfig.json
@@ -30,9 +30,6 @@
       "@models/*": [
         "./src/app/shared/models/*"
       ],
-      "@resolvers/*": [
-        "./src/app/shared/resolvers/*"
-      ],
       "@shared/*": [
         "./src/app/shared/*"
       ],
